requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.1
h11==0.14.0
idna==3.6
psycopg==3.1.18
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.29.0
